Working JetsProject

+++ src/com/skilldistillery/jets/entities/Flyable.java
@@ -1,6 +1,6 @@
 package com.skilldistillery.jets.entities;
 
 public interface Flyable {
-	double fly(int range, int speed);
-	double getSpeedInMach(int speed);
+	Double fly(Double range, Integer speed);
+	Double getSpeedInMach(Integer speed);
 }
